+++ src/Gaffer/FileSystemPath.cpp
@@ -54,8 +54,15 @@
 #ifndef _MSC_VER
 	#include <grp.h>
 	#include <pwd.h>
+#else
+	#include <stdio.h>
+	#include <Windows.h>
+	#include <tchar.h>
+	#include "accctrl.h"
+	#include "aclapi.h"
 #endif
 
+#include <regex>
 #include <sys/stat.h>
 
 using namespace std;
@@ -65,6 +72,143 @@ using namespace boost::posix_time;
 using namespace IECore;
 using namespace Gaffer;
 
+//////////////////////////////////////////////////////////////////////////
+// Internal utilities
+//////////////////////////////////////////////////////////////////////////
+
+namespace
+{
+
+#ifdef _MSC_VER
+
+std::string getFileSecurityInfo( const std::string &pathString, SECURITY_INFORMATION info )
+{
+	/// \todo : There may be optimizations to be had by caching credentials such
+	/// as the results of `LookupAccountSid` to avoid costly Windows API calls.
+	PSID pSidOwner = NULL;
+	PSID pSidGroup = NULL;
+	PSID pSid = NULL;
+	BOOL bRtnBool = TRUE;
+	LPTSTR AcctName = NULL;
+	LPTSTR DomainName = NULL;
+	DWORD dwAcctName = 1;
+	DWORD dwDomainName = 1;
+	SID_NAME_USE eUse = SidTypeUnknown;
+
+	DWORD result = GetNamedSecurityInfo(
+		pathString.c_str(),
+		SE_FILE_OBJECT,
+		OWNER_SECURITY_INFORMATION | GROUP_SECURITY_INFORMATION,
+		&pSidOwner,
+		&pSidGroup,
+		NULL,  // Out : DACL
+		NULL,  // Out : SACL
+		NULL  // Out : Security Descriptor
+	);
+
+	if( info == OWNER_SECURITY_INFORMATION )
+	{
+		pSid = pSidOwner;
+	}
+	else if( info == GROUP_SECURITY_INFORMATION )
+	{
+		pSid = pSidGroup;
+	}
+	else
+	{
+		return "";
+	}
+
+	// First call to LookupAccountSid to get the buffer sizes.
+	bRtnBool = LookupAccountSid(
+		NULL,
+		pSid,
+		AcctName,
+		( LPDWORD )&dwAcctName,
+		DomainName,
+		( LPDWORD )&dwDomainName,
+		&eUse
+	);
+
+	// Reallocate memory for the buffers.
+	AcctName = ( LPTSTR )GlobalAlloc( GMEM_FIXED, dwAcctName );
+
+	// Check GetLastError for GlobalAlloc error condition.
+	if (AcctName == NULL)
+	{
+		return "";
+	}
+
+	DomainName = ( LPTSTR )GlobalAlloc( GMEM_FIXED, dwDomainName );
+
+	// Check GetLastError for GlobalAlloc error condition.
+	if ( DomainName == NULL )
+	{
+		return "";
+	}
+
+	// Second call to LookupAccountSid to get the account name.
+	bRtnBool = LookupAccountSid(
+		NULL,
+		pSid,
+		AcctName,
+		(LPDWORD)&dwAcctName,
+		DomainName,
+		(LPDWORD)&dwDomainName,
+		&eUse
+	);
+
+	if ( bRtnBool == FALSE )
+	{
+		return "";
+	}
+
+	return AcctName;
+}
+
+#endif
+
+std::string getFileOwner( const std::string &pathString )
+{
+	std::string value;
+#ifndef _MSC_VER
+	struct stat s;
+	stat(pathString.c_str(), &s);
+	struct passwd *pw = getpwuid(s.st_uid);
+	return pw ? pw->pw_name : "";
+
+#else
+
+	return getFileSecurityInfo( pathString, OWNER_SECURITY_INFORMATION );
+
+#endif
+
+}
+
+std::string getFileGroup( const std::string &pathString )
+{
+	std::string value;
+#ifndef _MSC_VER
+	struct stat s;
+	stat(pathString.c_str(), &s);
+	struct group *gr = getgrgid( s.st_gid );
+
+	return gr ? gr->gr_name : "";
+
+#else
+
+	return getFileSecurityInfo( pathString, GROUP_SECURITY_INFORMATION );
+
+#endif
+
+}
+
+}  // namespace
+
+//////////////////////////////////////////////////////////////////////////
+// FileSystemPath implementation
+//////////////////////////////////////////////////////////////////////////
+
 IE_CORE_DEFINERUNTIMETYPED( FileSystemPath );
 
 static InternedString g_ownerPropertyName( "fileSystem:owner" );
@@ -73,14 +217,17 @@ static InternedString g_modificationTimePropertyName( "fileSystem:modificationTi
 static InternedString g_sizePropertyName( "fileSystem:size" );
 static InternedString g_frameRangePropertyName( "fileSystem:frameRange" );
 
+static std::regex g_driveLetterPattern{ "[A-Za-z]:" };
+
 FileSystemPath::FileSystemPath( PathFilterPtr filter, bool includeSequences )
 	:	Path( filter ), m_includeSequences( includeSequences )
 {
 }
 
 FileSystemPath::FileSystemPath( const std::string &path, PathFilterPtr filter, bool includeSequences )
-	:	Path( path, filter ), m_includeSequences( includeSequences )
+	:	Path( filter ), m_includeSequences( includeSequences )
 {
+	setFromString( path );
 }
 
 FileSystemPath::FileSystemPath( const Names &names, const IECore::InternedString &root, PathFilterPtr filter, bool includeSequences )
@@ -150,7 +297,7 @@ FileSequencePtr FileSystemPath::fileSequence() const
 
 	FileSequencePtr sequence = nullptr;
 	/// \todo Add cancellation support to `ls`.
-	IECore::ls( this->string(), sequence, /* minSequenceSize = */ 1 );
+	IECore::ls( this->nativeString(), sequence, /* minSequenceSize = */ 1 );
 	return sequence;
 }
 
@@ -189,14 +336,7 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
 					IECore::Canceller::check( canceller );
-					struct stat s;
-					stat( it->c_str(), &s );
-#ifndef _MSC_VER
-					struct passwd *pw = getpwuid( s.st_uid );
-					std::string value = pw ? pw->pw_name : "";
-#else
-					std::string value = "";
-#endif
+					std::string value = getFileOwner( it->c_str() );
 					std::pair<std::map<std::string,size_t>::iterator,bool> oIt = ownerCounter.insert( std::pair<std::string,size_t>( value, 0 ) );
 					oIt.first->second++;
 					if( oIt.first->second > maxCount )
@@ -210,14 +350,8 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 		}
 
 		std::string n = this->string();
-		struct stat s;
-		stat( n.c_str(), &s );
-#ifndef _MSC_VER
-		struct passwd *pw = getpwuid( s.st_uid );
-		return new StringData( pw ? pw->pw_name : "" );
-#else
-		return new StringData( "" );
-#endif
+
+		return new StringData( getFileOwner( n.c_str() ) );
 	}
 	else if( name == g_groupPropertyName )
 	{
@@ -237,14 +371,7 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
 					IECore::Canceller::check( canceller );
-					struct stat s;
-					stat( it->c_str(), &s );
-#ifndef _MSC_VER
-					struct group *gr = getgrgid( s.st_gid );
-					std::string value = gr ? gr->gr_name : "";
-#else
-					std::string value = "";
-#endif
+					std::string value = getFileGroup( *it );
 					std::pair<std::map<std::string,size_t>::iterator,bool> oIt = ownerCounter.insert( std::pair<std::string,size_t>( value, 0 ) );
 					oIt.first->second++;
 					if( oIt.first->second > maxCount )
@@ -258,14 +385,7 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 		}
 
 		std::string n = this->string();
-		struct stat s;
-		stat( n.c_str(), &s );
-#ifndef _MSC_VER
-		struct group *gr = getgrgid( s.st_gid );
-		return new StringData( gr ? gr->gr_name : "" );
-#else
-		return new StringData( "" );
-#endif
+		return new StringData( getFileGroup( n ) );
 	}
 	else if( name == g_modificationTimePropertyName )
 	{
@@ -369,7 +489,7 @@ void FileSystemPath::doChildren( std::vector<PathPtr> &children, const IECore::C
 	{
 		IECore::Canceller::check( canceller );
 		std::vector<FileSequencePtr> sequences;
-		IECore::ls( p.string(), sequences, /* minSequenceSize */ 1 );
+		IECore::ls( this->nativeString(), sequences, /* minSequenceSize */ 1 );
 		for( std::vector<FileSequencePtr>::iterator it = sequences.begin(); it != sequences.end(); ++it )
 		{
 			IECore::Canceller::check( canceller );
@@ -450,4 +570,64 @@ PathFilterPtr FileSystemPath::createStandardFilter( const std::vector<std::strin
 	result->addFilter( searchFilter );
 
 	return result;
+}
+
+#ifdef _MSC_VER
+
+void FileSystemPath::rootAndNames(const std::string &string, InternedString &root, Names &names ) const
+{
+	std::string sanitizedString = string;
+
+	// If `string` is coming from a PathMatcher, it will always have a single leading slash.
+	// On Windows, check to see if the first element is a drive letter path and strip the leading
+	// slash if so.
+	if( sanitizedString.size() && sanitizedString[0] == '/' )
+	{
+		Names splitPath;
+		StringAlgo::tokenize(sanitizedString, '/', splitPath);
+		if( splitPath.size() )
+		{
+			const std::string firstElement = splitPath[0].string();
+			if( std::regex_match( firstElement, g_driveLetterPattern ) )
+			{
+				sanitizedString.erase( sanitizedString.begin(), sanitizedString.begin() + 1 );
+			}
+		}
+	}
+
+	const path convertedPath( sanitizedString );
+	root = convertedPath.root_path().generic_string();
+
+	path::const_iterator startIt = convertedPath.begin();
+
+	// path iteration includes the root name and directory, if present
+	if( convertedPath.has_root_name() )
+	{
+		++startIt;
+	}
+
+	if( convertedPath.has_root_directory() )
+	{
+		++startIt;
+	}
+
+	for( path::const_iterator it = startIt, eIt = convertedPath.end(); it != eIt; ++it )
+	{
+		names.push_back( it->string() );
+	}
+}
+
+#endif
+
+std::string FileSystemPath::nativeString() const
+{
+#ifndef _MSC_VER
+	return string();
+#endif
+
+	path p( string() );
+	// This is used instead of `nativeString()` because `nativeString()` on Windows
+	// returns a `wstring`.
+	p.make_preferred();
+	return p.string();
 }